refactor: Final cleanup of Repository
Got rid of static model project and renamed remaining files accordingly
