+++ ExcelReader.BrozDa/ExcelReader.BrozDa.sln
@@ -1,11 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.12.35514.174 d17.12
+VisualStudioVersion = 17.12.35514.174
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExcelReader.BrozDa", "ExcelReader.BrozDa\ExcelReader.BrozDa.csproj", "{5410F29A-A411-421B-911B-A8376E4838B7}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExcelReaderDynamic", "ExcelReaderDynamic\ExcelReaderDynamic.csproj", "{B5CE63A6-2E46-448D-8E2B-AA5311429049}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExcelReader.Brozda", "ExcelReader.Brozda\ExcelReader.Brozda.csproj", "{B5CE63A6-2E46-448D-8E2B-AA5311429049}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -13,10 +11,6 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{5410F29A-A411-421B-911B-A8376E4838B7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5410F29A-A411-421B-911B-A8376E4838B7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5410F29A-A411-421B-911B-A8376E4838B7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5410F29A-A411-421B-911B-A8376E4838B7}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B5CE63A6-2E46-448D-8E2B-AA5311429049}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B5CE63A6-2E46-448D-8E2B-AA5311429049}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B5CE63A6-2E46-448D-8E2B-AA5311429049}.Release|Any CPU.ActiveCfg = Release|Any CPU
