--- ExcelReader.BrozDa/ExcelReader.BrozDa/Data/ExcelReaderContext.cs
@@ -1,31 +1,0 @@
-﻿using ExcelReader.BrozDa.Models;
-using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ExcelReader.BrozDa.Data
-{
-    internal class ExcelReaderContext : DbContext
-    {
-        private string _connectionString = @"Data Source=(localdb)\LOCALDB;Initial Catalog=ExcelReader;Integrated Security=True;Connect Timeout=30;Encrypt=True;Trust Server Certificate=False;Application Intent=ReadWrite;Multi Subnet Failover=False";
-
-        DbSet<Person> Persons { get; set; } = null!;
-
-        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        {
-            optionsBuilder.UseSqlServer(_connectionString);
-        }
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-           
-            modelBuilder.Entity<Person>()
-                .HasKey(p => p.Id);
-            modelBuilder.Entity<Person>()
-                .Property(p => p.UserId)
-                .IsRequired();
-        }
-    }
-}
