--- ExcelReader.BrozDa/ExcelReader.BrozDa/Models/Person.cs
@@ -1,13 +1,0 @@
-﻿namespace ExcelReader.BrozDa.Models
-{
-    internal class Person
-    {
-        public int Id { get; set; }
-        public string UserId { get; set; } = null!;
-        public string FirstName { get; set; } = null!;
-        public string LastName { get; set; } = null!;
-        public string Email { get; set; } = null!;
-        public string Phone { get; set; } = null!;
-        public DateTime? BirthDate { get; set; }
-    }
-}
