+++ ExcelReader.BrozDa/ExcelReader.BrozDa/Models/Record.cs
