--- ExcelReader.BrozDa/ExcelReader.BrozDa/Properties/launchSettings.json
@@ -1,8 +1,0 @@
-{
-  "profiles": {
-    "ExcelReader.BrozDa": {
-      "commandName": "Project",
-      "workingDirectory": "E:\\Git Repos\\CodeReviews.Console.ExcelReader\\ExcelReader.BrozDa\\ExcelReader.BrozDa"
-    }
-  }
-}
