+++ ExcelReader.BrozDa/ExcelReader.BrozDa/appconfig.json
@@ -1,7 +1,6 @@
 {
-  "EPPlus": {
-    "ExcelPackage": {
-      "LicenseContext": "NonCommercialPersonal:BrozDa"
-    }
+  "ConnectionStrings": {
+    "ExcelReaderDB": "Data Source=(localdb)\\LOCALDB;Initial Catalog=ExcelReaderDynamic;Integrated Security=True;",
+    "MasterDB": "Data Source=(localdb)\\LOCALDB;Initial Catalog=master;Integrated Security=True;"
   }
 }
